+++ build_win_exe.bat
@@ -1,1 +1,1 @@
-pyinstaller jupyter_notebook_new.spec --onefile --icon=app.ico
+pyinstaller jupyter_notebook.spec --clean --onefile --icon=app.ico
